src/inference/requirements.txt
torch
transformers
timm
mediapipe==0.8.3.1
tqdm
scikit-learn
xgboost ## installs CPU version, check docs for GPU compatible version
uvicorn==0.13.4
fastapi==0.65.1
python-dotenv==0.17.1
azure-storage-blob==12.8.1